File rename: nwpu.py -> vhr10.py
