+++ tests/data/vhr10/NWPU VHR-10 dataset.rar
